Web bridge refactoring
